Move index creation to hana.go so it can be used as stand alone functions
