+++ internal/backends/hana/collection.go
@@ -18,8 +18,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"sort"
-	"strings"
 
 	"github.com/SAP/go-hdb/driver"
 	"github.com/google/uuid"
@@ -363,138 +361,14 @@ func (c *collection) Compact(ctx context.Context, params *backends.CompactParams
 	return new(backends.CompactResult), nil
 }
 
-// Prefixes an index with the collection name to store it in hana.
-//
-// Reasoning:
-// Hana DocStore stores indexes on a schema/database level, that may cause
-// confilts for indexes of different collections having the same name (eg col1.idx and col2.idx).
-func (c *collection) prefixIndexName(indexName string) string {
-	return fmt.Sprintf("%s__%s", c.name, indexName)
-}
-
-// Removes the prefix of an index name.
-func (c *collection) removeIndexNamePrefix(prefixedIndex string) string {
-	prefix := fmt.Sprintf("%s__", c.name)
-	return strings.Replace(prefixedIndex, prefix, "", 1)
-}
-
 // ListIndexes implements backends.Collection interface.
 func (c *collection) ListIndexes(ctx context.Context, params *backends.ListIndexesParams) (*backends.ListIndexesResult, error) {
-	db, err := databaseExists(ctx, c.hdb, c.database)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	if !db {
-		return nil, backends.NewError(
-			backends.ErrorCodeDatabaseDoesNotExist,
-			lazyerrors.Errorf("no ns %s.%s", c.database, c.name),
-		)
-	}
-
-	col, err := collectionExists(ctx, c.hdb, c.database, c.name)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	if !col {
-		return nil, backends.NewError(
-			backends.ErrorCodeCollectionDoesNotExist,
-			lazyerrors.Errorf("no ns %s.%s", c.database, c.name),
-		)
-	}
-
-	sql := "SELECT idx.INDEX_NAME, idx.COLUMN_NAME, idx.ASCENDING_ORDER " +
-		"FROM INDEX_COLUMNS idx, M_TABLES tbl " +
-		"WHERE idx.SCHEMA_NAME = '%s' AND idx.TABLE_NAME = '%s' AND " +
-		"idx.SCHEMA_NAME = tbl.SCHEMA_NAME AND idx.TABLE_NAME = tbl.TABLE_NAME AND " +
-		"tbl.TABLE_TYPE = 'COLLECTION'"
-
-	sql = fmt.Sprintf(sql, c.database, c.name)
-
-	rows, err := c.hdb.QueryContext(ctx, sql)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	defer rows.Close()
-
-	var indexes []backends.IndexInfo
-
-	for rows.Next() {
-		var idxName, idxColName string
-		var ascending bool
-
-		err = rows.Scan(&idxName, &idxColName, &ascending)
-		if err != nil {
-			return nil, lazyerrors.Error(nil)
-		}
-
-		indexes = append(indexes, backends.IndexInfo{
-			Name:   c.removeIndexNamePrefix(idxName),
-			Unique: true, // HANATODO: Is this possible to query from HANA?
-			Key:    []backends.IndexKeyPair{{Field: idxColName, Descending: !ascending}},
-		})
-	}
-
-	res := backends.ListIndexesResult{Indexes: indexes}
-
-	sort.Slice(res.Indexes, func(i, j int) bool {
-		return res.Indexes[i].Name < res.Indexes[j].Name
-	})
-
-	return &res, nil
-}
-
-// indexExists checks if an index exists in a list of indexes retrieved from HANA.
-func indexExists(indexes []backends.IndexInfo, indexToFind string) bool {
-	for _, idx := range indexes {
-		if idx.Name == indexToFind {
-			return true
-		}
-	}
-
-	return false
+	return listIndexes(ctx, c.hdb, c.database, c.name)
 }
 
 // CreateIndexes implements backends.Collection interface.
 func (c *collection) CreateIndexes(ctx context.Context, params *backends.CreateIndexesParams) (*backends.CreateIndexesResult, error) { //nolint:lll // for readability
-	db, err := databaseExists(ctx, c.hdb, c.database)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	if !db {
-		panic("database does not exist")
-	}
-
-	err = createCollectionIfNotExists(ctx, c.hdb, c.database, c.name)
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	existingIndexes, err := c.ListIndexes(ctx, new(backends.ListIndexesParams))
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	sql := "CREATE HASH INDEX %q.%q ON %q.%q(%q)"
-
-	var createStmt string
-
-	// HANATODO Can we support more than one field for indexes in HANA DocStore?
-	for _, index := range params.Indexes {
-		if !indexExists(existingIndexes.Indexes, index.Name) {
-			createStmt = fmt.Sprintf(sql, c.database, c.prefixIndexName(index.Name), c.database, c.name, index.Key[0].Field)
-
-			_, err := c.hdb.ExecContext(ctx, createStmt)
-			if err != nil {
-				return nil, lazyerrors.Error(err)
-			}
-		}
-	}
-
-	return new(backends.CreateIndexesResult), nil
+	return createIndexes(ctx, c.hdb, c.database, c.name, params)
 }
 
 // DropIndexes implements backends.Collection interface.
@@ -511,7 +385,7 @@ func (c *collection) DropIndexes(ctx context.Context, params *backends.DropIndex
 	var droptStmt string
 
 	for _, index := range params.Indexes {
-		droptStmt = fmt.Sprintf(sql, c.database, c.prefixIndexName(index))
+		droptStmt = fmt.Sprintf(sql, c.database, prefixIndexName(c.name, index))
 
 		_, err := c.hdb.ExecContext(ctx, droptStmt)
 		if err != nil {
